pyproject.toml
[tool.poetry]
readme = "README.md"
packages = [
    { include = "ibanfirst_client", from = "src" },
    { include = "orness"}
]
[project]
name = "OrnessSDK"
version = "1.0.0"
authors = [{name = "randy BEKALE", email = "randy.bekale@orness.com"}, ]
dynamic = ['readme']
description = ""

[tool.poetry.dependencies]
python = "^3.10"
certifi = "^2025.1.31"
six = "^1.17.0"
python-dateutil = "^2.9.0.post0"
setuptools = "^78.1.0"
urllib3 = "^2.3.0"
redis = "^5.2.1"
python-dotenv = "^1.1.0"
jsonschema = "^4.23.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
country-converter = "^1.3"
pyyaml = "^6.0.2"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


